feat: Update styles and interaction logic in Wallet and Notifications modules
Changes were made in several files to improve the UI and usability of the Wallet and Notification modules. The styles of the connect wallet view and notification button were updated. Logic in the notification controller was modified to better manage notification count data. The interaction handling of the wallet view in mobile screen size was improved by properly showing the connect view. Furthermore, disabled state was added to notification open button when there are no notifications.