feat: Refactor code for better imports and update service worker registration
This commit has improved the way modules are imported across the project for better readability and getting rid of the relative path issues. It also includes changes in the service worker registration by logging worker updates, promising the worker update and removing an unnecessary space in `storage-controller.ts`. In addition, the PWA register in vite has been updated to skip waiting and claim clients.